requirements.txt
# Core dependencies
git+https://github.com/MatthewScholefield/pocketsphinx-python@patch-2#egg=pocketsphinx

PyYAML
pyaml
Twiggy
requests
# pocketsphinx
SpeechRecognition
PyAudio
pyserial
pyalsaaudio
tornado
websocket-client
pyinotify

fann2==1.0.7
padatious

# Skill dependencies
multi-key-dict
pyowm
ddg3
parsedatetime
geopy
pytz
pyjokes
cleverwrap
pydora
psutil
py-stackexchange
lxml
